requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
bert-serving-client==1.9.1
bert-serving-server==1.9.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cx-Oracle==7.1.3
cycler==0.10.0
docker==4.0.1
GPUtil==1.4.0
idna==2.8
importlib-metadata==0.18
isort==4.3.20
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.4
matplotlib==3.1.0
mccabe==0.6.1
more-itertools==7.0.0
numpy==1.16.4
packaging==19.0
pluggy==0.12.0
py==1.8.0
pylint==2.3.1
pyparsing==2.4.0
pypiwin32==223
pytest==4.6.3
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.3
pywin32==224
pyzmq==18.0.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.4
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
-e git+https://github.com/bradfox2/WordParsing@f8aca14138d2668f0254feec62a73565d9132371#egg=WordParsing
wrapt==1.11.1
zipp==0.5.1
